refactor update course

+++ app/infrastructure/course/course_repository.py
@@ -47,13 +47,18 @@ class CourseRepositoryImpl(CourseRepository):
         course_dto = CourseDTO.from_entity(course)
         try:
             _course = self.session.query(CourseDTO).filter_by(id=course_dto.id).one()
-            _course.name = course_dto.name
-            _course.language = course_dto.language
-            _course.price = course_dto.price
-            _course.description = course_dto.description
+            if course_dto.name:
+                _course.name = course_dto.name
+            if course_dto.language:
+                _course.language = course_dto.language
+            if course_dto.price is not None:
+                _course.price = course_dto.price
+            if course_dto.description:
+                _course.description = course_dto.description
             _course.updated_at = course_dto.updated_at
-            self.session.query(Category).filter_by(course_id=course_dto.id).delete()
-            _course.categories = course_dto.categories
+            if course_dto.categories:
+                self.session.query(Category).filter_by(course_id=course_dto.id).delete()
+                _course.categories = course_dto.categories
         except:
             raise
